+++ crates/stackable-versioned-macros/src/lib.rs
@@ -939,16 +939,16 @@ where
 }
 
 #[cfg(test)]
-mod test {
+mod snapshot_tests {
     use insta::{assert_snapshot, glob};
 
     use super::*;
 
     #[test]
-    fn default_snapshots() {
+    fn default() {
         let _settings_guard = test_utils::set_snapshot_path().bind_to_scope();
 
-        glob!("../fixtures/inputs/default", "*.rs", |path| {
+        glob!("../tests/inputs/default/pass", "*.rs", |path| {
             let formatted = test_utils::expand_from_file(path)
                 .inspect_err(|err| eprintln!("{err}"))
                 .unwrap();
@@ -958,10 +958,10 @@ mod test {
 
     #[cfg(feature = "k8s")]
     #[test]
-    fn k8s_snapshots() {
+    fn k8s() {
         let _settings_guard = test_utils::set_snapshot_path().bind_to_scope();
 
-        glob!("../fixtures/inputs/k8s", "*.rs", |path| {
+        glob!("../tests/inputs/k8s/pass", "*.rs", |path| {
             let formatted = test_utils::expand_from_file(path)
                 .inspect_err(|err| eprintln!("{err}"))
                 .unwrap();