+++ crates/stackable-versioned/CHANGELOG.md
@@ -13,17 +13,20 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Add `kube_client` crate override to `k8s(crates())` to specify a custom import path. This override
+  will not be passed to the `#[kube()]` attribute, but will only be available to internal
+  `#[versioned]` macro code ([#1038]).
 - BREAKING: The `convert_with` parameter of the `changed()` action was renamed and split into two
   parts to be able to control the conversion during upgrades and downgrades: `upgrade_with` and
   `downgrade_with` ([#1033]).
-
-[#1033]: https://github.com/stackabletech/operator-rs/pull/1033
 
 ### Fixed
 
 - Correctly handle fields added in later versions ([#1031]).
 
 [#1031]: https://github.com/stackabletech/operator-rs/pull/1031
+[#1033]: https://github.com/stackabletech/operator-rs/pull/1033
+[#1038]: https://github.com/stackabletech/operator-rs/pull/1038
 
 ## [0.7.1] - 2025-04-02